update AM contract tests
